requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.3
django-cors-headers==3.10.0
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Pillow==8.4.0
psycopg2==2.9.3
pyrsistent==0.19.3
python-dotenv==0.20.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
